[2.3.0] AbstractHealthConfiguration introduced to simplify same configs ArangoConfigurationTests improved
