+++ src/main/java/io/micronaut/configuration/arango/health/AbstractHealthConfiguration.java
@@ -1,0 +1,49 @@
+package io.micronaut.configuration.arango.health;
+
+import io.micronaut.context.exceptions.ConfigurationException;
+
+/**
+ * @author Anton Kurako (GoodforGod)
+ * @since 13.08.2021
+ */
+public abstract class AbstractHealthConfiguration {
+
+    private boolean enabled = true;
+    private long timeoutInMillis = 5000;
+    private int retry = 2;
+
+    public boolean isEnabled() {
+        return enabled;
+    }
+
+    public void setEnabled(boolean enabled) {
+        this.enabled = enabled;
+    }
+
+    public long getTimeoutInMillis() {
+        return timeoutInMillis;
+    }
+
+    public void setTimeoutInMillis(long timeoutInMillis) {
+        if (timeoutInMillis < 0)
+            throw new ConfigurationException("Timeout for health can not be less than 0");
+        this.timeoutInMillis = timeoutInMillis;
+    }
+
+    public int getRetry() {
+        return retry;
+    }
+
+    public void setRetry(int retry) {
+        if (retry < 1)
+            throw new ConfigurationException("Retry for health can not be less than 1");
+        this.retry = retry;
+    }
+
+    @Override
+    public String toString() {
+        return "[enabled=" + enabled +
+                ", timeoutInMillis=" + timeoutInMillis +
+                ", retry=" + retry + ']';
+    }
+}
